refactor: Get tests passing by guariding against a polyfil bug.
Don't use ShadyCSS polyfill. Re-organise tests to reflect deprecations.
